traceapp: add support for clicking and hovering on timespan labels
Before this change, clicking or hovering over a timespan's label will
not work. If the text width is almost equal to the timespan's rectangle
width, or if the rectangle width is very small, it is very difficult to
click or hover over it (as the text blocks events from going to the
rectangle).

With this change, it is possible to hover over (to update the timeline
selector, or "hoverRes" div, in the bottom-left corner of the timeline)
and it is also possible to click on text label of a timespan to view
it's sub-spans.

From what I have found, d3-timeline does not itself support click or
hover events on timespan labels. To solve this, I've employed a workaround
using JQuery by selecting the next element and parsing the ID of it (which
has a data index suffix, i.e. the information associated with the timespan
rectangle).