traceapp: Fix display of traces that contain no TimespanEvent's
When a *apptrace.Trace contains no TimespanEvent's it causes the JSON
translation to have no time set information (see #6), and causes
d3-timeline.js to crash in an attempt to iterates over the null array
object.

As a solution, we return a nil dataset in this case. We also only
initialize the d3 timeline in the event that we have a non-null dataset,
which allows JavaScript to continue execution and gives the page a
cleaner display overall.

Helps #6