traceapp: enable clicking on profile-view table rows (go to span page).
This change enables clicking on table rows under profile-view. Clicking
any row brings you to the span/sub-span page for that trace, as it would
if you clicked it in the timeline. This allows one to find both the span
ID as well as identify what span it is when e.g. names are the same or
otherwise non-helpful.

Fixes #25