traceapp: generate profile times that exactly match the timeline.
This change makes the profiling times be calculated as floating-point
numbers and rounded up, as they are in the d3 timeline. This is both
more accurate and less confusing to users overall.

This change also appropriately grabs the largest timespan events, rather
than just the first one. This matches how apptrace already sends the
timespans to the template for rendering in the timeline.