Improve RecentStore.Collect performance by ~99%.
There is a guard to protect against the invokation of RecentStore's
evictBefore method, which looks like this:

  if time.Since(rs.lastEvicted) > rs.MinEvictAge {
      rs.evictBefore(time.Now().Add(-1 * rs.MinEvictAge))
  }

Effectively, if the time since the last eviction is less than the minimum
eviction age, another eviction does not need to occur.

evictBefore has two primary costs to it -- one, it iterates over each item
in the store to see if it needs to be evicted, and two, it spawns a goroutine
to later delete the items-to-be-evicted. These costs are fine, as long as it's
not invoked too often.

As it turns out rs.lastEvicted is never set, though. For each call to Collect
on a RecentStore it is invoked, and subsequently a linear search and the
spawning of a goroutine.

For BenchmarkRecentStore500, the total number of evictBefore calls lowers from
50500 to 3, and appropriately the time spent also lowers dramatically:

benchmark                   old ns/op     new ns/op     delta
BenchmarkRecentStore500     483523346     4424602       -99.08%