traceapp: add a link in trace view to copy as JSON (trace exporting).
JSON is human-readable and thus preferred to a binary format -- someone
giving their trace to someone else will be able to directly see what is
inside of it. One downside is that all annotation values are arbitrary
byte sequences -- and are thus encoded into JSON as base64, though it's
probably not that big of a deal as all the info can be viewed through
the UI directly instead.

HTML5 does not have a copy-to-clipboard API and thus Flash must be used
to implement this feature (suprisingly). A generally accepted solution
is to make use of ZeroClipboard, a small JS/Flash library that does it.