Enable using LimitStore/RecentStore functionality with AggregateStore
This change enables the usage of LimitStore / RecentStore / etc functionality
in tangent with AggregateStore. For example I want to keep the N-slowest
traces and aggregated trace data (what AggregateStore provides) plus also keep
the last 5 minutes of traces (regardless of whether or not they are the
N-slowest).

Prior to this change it was not possible. You could send collections to both
stores, but if a trace happended to be one of the ones in the last 5 minutes
and also one of the N-slowest you would end up with duplicate data in the
backing store (assuming they went to the same backing store).