Implement a workaround to avoid a memory leak in AggregateStore.
as.groups will have a single ID key placed into the map upon each collection
(e.g. HTTP request) and doing that needs the storage of 8 bytes of data (traceID
and pointer). Prior to this change as.groups only had items removed upon the
deletion of an entire group (e.g. if a HTTP route had not been accessed _at all_
over the past 72/hrs) so it would leak memory.

Removing items from as.groups is non-trivial because there is no clear-cut place
at which items in the map are no longer needed. As a workaround for now we employ
a timer of 30s, after which we suggest that items in as.groups are no longer
needed. In the unlikely case they are needed after 30s, that annotation would go
missing (would not be associated with the trace).