Improve performance / lessen garbage generated by ChunkedCollector.
This change improves the performance of ChunkedCollector by having it not
perform needless conversion between annotations and their (protobuf) wire
format. This improves the performance of it significantly and will also likely
reduce GC times due to the fact that the Protobuf wire format structures are
largely pointer-based.

When using:

  go test -memprofile=mem.out -memprofilerate=1 -benchtime=15s -bench="BenchmarkChunkedCollector" .

benchcmp shows:

  benchmark                          old ns/op     new ns/op     delta
  BenchmarkChunkedCollector500-4     3715457       363884        -90.21%

Not shown by these benchmarks is the fact that GC pauses will likely be less
significant after this change as well.