Improve storage model / performance of AggregateStore.
This is a breaking change that improves the storage model and performance
of AggregateStore significantly. Prior to this change all Timespans for a
given spanGroup would be stored as part of the AggregateEvent as a simple
slice. In practice the old AggregateEvent must be removed from the
MemoryStore and the new/updated one inserted once again. However doing so
requires marshaling and unmarshaling the AggregateEvent, which makes the
insertion of a single timespan for a spanGroup an O(n) operation. To solve
this we keep AggregateEvent simple and move all timespans out of it into
their own subspans. This matches the Appdash storage model much better and
(most importantly) makes insertion of a single timespan event an O(1)
operation (aside from whatever the storage backend might have to perform).

Additionally, we update the Trace.Aggregated function to return relevant
information:

  func (t *Trace) Aggregated() (*AggregateEvent, []TimespanEvent, error)

Note the transition to `*AggregateEvent` instead of `[]AggregateEvent`: we
never actually had a scenario in which multiple AggregateEvents would
appear anyway, so this is a simplification of the API.