replace AggregateStore.Keep with MultiStore and MultiQueryer
AggregateStore.Keep is hard to reason about because it makes the assumption
that the Keep store will be responsible for removing traces from the output
store in the future.

The initial idea with Keep was to have a single backing MemoryStore for both
a LimitStore and AggregateStore, so once a trace was evicted from the LimitStore
it would be subsequently removed from the backing MemoryStore. However in
practice the two cannot share the same backing store as it produces an implicit
memory leak (AggregateStore will ask LimitStore "should we keep this trace?"
and it will respond "yep, we have that one, keep it!" even though LimitStore
wasn't the one that added it to the backing MemoryStore (AggregateStore did!).

Since the duplication between AggregateStore and LimitStore is very rare / low
(at max the N-slowest traces would also be in LimitStore), it's much more sane
to have each backed by their own MemoryStore and make use of multiple-insertion
(MultiStore) and multiple-query (MultiQueryer) as the flow-control points in
your app.

This change removes AggregateStore.Keep and adds MultiStore and MultiQueryer
functionality.