Refactor template data file structure to be more like appdash-data repo.
Move template files into data subdirectory (which contains only the
template files).

Rename `Assets` variable to `Data`.

Remove relative path code from data_rootdir.go. I think the reasons it
was added in the past are no longer applicable and it shouldn't be
needed now. If this breaks something, please open an issue and we can
investigate and revert this change if needed.