httptrace: Collect Name & User after inner handler
This allows inner handlers to run and set information which the MiddlewareConfig
may want to use. A concrete example of this is having the middleware run before
mux has had a chance to set a route name.

This could potentially break existing uses if the inner handlers destroy or
mutate the information that MiddlewareConfig depended on.