InfluxDBStore: add InfluxDBStoreConfig.LogOutput for configurable log output
This change makes it such that InfluxDBStore can have it's log output
location directly specified by the end user. This means it can go to stderr,
stdout, or even a file for example.

This is a very important change for Appdash users, because it means that their
InfluxDB logs, which are only indirectly related to their application are no
longer polluting their application logs (users embed Appdash, which in turn
embeds InfluxDB, which then proceeds to pollute their apps logs).