Replace gorilla/context with stdlib context
Negroni started using context.Context to manage internal data, and
http.Request.WithContext() to set it, which means that the pointer
addresses are no longer the same between the gorilla/context.Set() and
Get() calls, causing nil to be returned from Get().

This change switches over to using the standard library context
implementation to propagate the span ID from the middleware to the
handler.